# :zap: make just two parallel requests to teamtailor