When deployed to cloudflare this speeds up by ~200ms. And in practice even more, it seems the api gets faster when making frequent requests, but cold starts are much slower. The worst cases with current implementation is around 1-2 seconds
<img width="79" alt="image" src="https://user-images.githubusercontent.com/244257/197177061-cc496917-7f28-4e71-88a1-3da42e386f5e.png">


| Before  | After  |
|---|---|
| <img width="56" alt="image" src="https://user-images.githubusercontent.com/244257/197176350-3f7b90a8-5f51-4703-b7d0-291154c42cdf.png"> | <img width="68" alt="image" src="https://user-images.githubusercontent.com/244257/197176282-091fee47-6b9f-4c4e-934f-a19ecd9f8dd3.png"> |